federation-api: Add query server key discovery endpoints
