+++ ruma-federation-api/src/discovery/get_remote_server_keys/mod.rs
@@ -1,0 +1,3 @@
+//! Query for another server's keys. The receiving (notary) server must sign the keys returned by the queried server.
+
+pub mod v2;
