+++ ruma-federation-api/src/discovery/get_remote_server_keys_batch/mod.rs
@@ -1,0 +1,3 @@
+//! Query for keys from multiple servers in a batch format. The receiving (notary) server must sign the keys returned by the queried servers.
+
+pub mod v2;
